// src/api/generated/models/postCompaniesCompanyIdImportsRequestLinesCsvPreview200.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { PostCompaniesCompanyIdImportsRequestLinesCsvPreview200Preview } from './postCompaniesCompanyIdImportsRequestLinesCsvPreview200Preview';
import type { PostCompaniesCompanyIdImportsRequestLinesCsvPreview200ColumnMapping } from './postCompaniesCompanyIdImportsRequestLinesCsvPreview200ColumnMapping';

export type PostCompaniesCompanyIdImportsRequestLinesCsvPreview200 = {
  preview: PostCompaniesCompanyIdImportsRequestLinesCsvPreview200Preview;
  columnMapping: PostCompaniesCompanyIdImportsRequestLinesCsvPreview200ColumnMapping;
};